fix: Spring AI 1.0.0 GA API compatibility & code quality improvements
- Fix AnswerRelevancyEvaluator: embed(List<String>) returns List<float[]> not float[][]
- Fix DataAnalysisAgent: null-safe JDBC column access (NPE prevention)
- Fix PublishController: Math.abs(Integer.MIN_VALUE) boundary bug
- Fix ContentController/TopicController: null-safe query string building
- Fix RRFFusionStrategy: add null element defense in accumulate()
- Fix JacksonConfig: remove duplicate ObjectMapper bean (customizer not applied)
- Fix GlobalExceptionHandler: HTTP status code range validation
- Remove duplicate StructuredOutputException, RagasEvaluation, RagasReport
- Add StructuredOutputException handler (HTTP 422)
- All 206 tests pass, 0 failures

+++ src/main/java/com/contentops/ai/api/controller/PublishController.java
@@ -43,7 +43,7 @@ public class PublishController {
         params.put("platforms", request.getPlatforms());
 
         AgentRequest agentRequest = AgentRequest.builder()
-                .query("publish:" + request.getPlatforms() + ":" + Math.abs(request.getContent().hashCode()))
+                .query("publish:" + request.getPlatforms() + ":" + Math.abs((long) request.getContent().hashCode()))
                 .tenantId(TenantContext.resolveTenantCode(tenantId))
                 .params(params)
                 .build();
